requirements.txt
fastapi
uvicorn
langchain-community
nltk
wikipedia
pydantic
typing-extensions
huggingface-hub
transformers
sentence-transformers
aiofiles
slowapi
firebase-admin
apscheduler
httpx
google-auth
requests
python-dotenv
redis
google-cloud-aiplatform
pinecone-client
google-cloud-bigquery
google-cloud-storage
diffusers
pinecone
python-multipart
google